Save Point #1

+++ WebApplication2/Models/BusinessPartner.cs
@@ -5,7 +5,7 @@
         public int BusinessPartnerId { get; set; }
         public string? Role { get; set; } = string.Empty;
         public string? Fullname { get; set; } = string.Empty;
-        public int? CustomerCode { get; set; }
+        public long? CustomerCode { get; set; }
         public List<SalesTransaction>? SalesTransaction { get; set; }
     }
 }
