+++ WebApplication2/Models/SalesTransaction.cs
@@ -5,7 +5,7 @@ namespace WebApplication2.Models
     public class SalesTransaction
     {
         public int SalesTransactionId { get; set; }
-        public int ContractNumber { get; set; }
+        public long ContractNumber { get; set; }
         public string? TypeOfSale { get; set; } = string.Empty;
         public int? BusinessPartnerId { get; set; }
         public int? PropertyId { get; set; }
